P3/backend/requirements.txt

Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7